main/rag/requirements.txt
fastapi
python-multipart
uvicorn
sentence-transformers
qdrant-client
transformers
python-dotenv